fix test import
